Adjusted potentiometer and limit switches to read more intuitively
Changed potentiometer full range and off-set so the value will increase
when the POT is turned clockwise and decrease when it is turned
counter-clockwise.
Change the wiring on the limit switches to read "NORMAL CLOSED" or "NC".
Then adjusted the code to match the wiring.
Also changed the channel number for the camera switch on the dashboard
to a working switch on the dashboard (last year's grabber open/close
switch  on channel 4).